requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
asyncpg==0.29.0
openpyxl==3.1.2
pandas==2.1.0
